functions/google-vision/requirements.txt
boto3
google-cloud-vision
pillow
